Add warning about first four lines.
Beginning to take the stance that they *must* be the first four lines
in a post file, even if not required by everything yet. It makes things
eaiser to have this assumption.

Currently I think only the modification date is required to be on the
second line due to the replacement process after editing. Everything else
could tecnically have comment lines inbetween them.

At this point, I'm deciding to disallow that.