Don't make /pubkey.gpg a 404 if not signing pages ...
Also clean up old signature files if we just
swithced to not signing pages.

This commit also adds a make target for "simple actions" that should only be
executed once when a config file changes. For now at least, it only really makes
sense to put cleanup type commands in it.

Closes #97